% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_all_tools.R
\name{has_all_tools}
\alias{has_all_tools}
\title{General install test}
\usage{
has_all_tools(debug = FALSE)
}
\arguments{
\item{debug}{if \code{TRUE}, will print out extra information for debugging.
Defaults to `FALSE`.}
}
\value{
a single logical value
}
\description{
This function just checks whether you have the necessary packages for
following the course.
}
